# docs/api/instainstru-openapi.yaml
openapi: 3.0.3
info:
  title: InstaInstru API
  description: |
    Backend API for InstaInstru - A platform connecting students with instructors for instant bookings.

    The InstaInstru API provides endpoints for user authentication, instructor profile management,
    availability scheduling, and instant booking of lessons. All endpoints use JSON for requests
    and responses, with JWT Bearer token authentication.
  version: 1.0.0
  contact:
    email: support@instainstru.com
  license:
    name: Proprietary
servers:
  - url: https://api.instainstru.com
    description: Production server
  - url: http://localhost:8000
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /instructors/:
    get:
      tags:
        - instructors
      summary: List all instructors
      description: Get a list of all instructor profiles with their services
      operationId: listInstructors
      security: []
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
          description: Number of records to skip
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            maximum: 100
          description: Maximum number of records to return
      responses:
        '200':
          description: List of instructor profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstructorProfileResponse'
...
  /instructors/{instructor_id}:
    get:
      tags:
        - instructors
      summary: Get instructor by ID
      description: Get a specific instructor's profile by user ID
      operationId: getInstructor
      security: []
      parameters:
        - in: path
          name: instructor_id
          required: true
          schema:
            type: integer
          description: Instructor's user ID
      responses:
        '200':
          description: Instructor profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstructorProfileResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Availability Management
...
  /instructors/bookings/pending-completion:
    get:
      tags:
        - instructor-bookings
      summary: Get pending lessons
      description: Get lessons awaiting completion by instructor
      operationId: getPendingLessons
      responses:
        '200':
          description: List of pending lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /instructors/bookings/completed:
    get:
      tags:
        - instructor-bookings
      summary: Get completed lessons
      description: Get instructor's completed lessons history
      operationId: getCompletedLessons
      responses:
        '200':
          description: List of completed lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /instructors/bookings/{booking_id}/complete:
    post:
      tags:
        - instructor-bookings
      summary: Complete lesson
      description: Mark a lesson as completed by instructor
      operationId: completeLessonByInstructor
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: string
          description: Booking ID to complete
      responses:
        '200':
          description: Lesson marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /instructors/bookings/{booking_id}/dispute:
    post:
      tags:
        - instructor-bookings
      summary: Dispute completion
      description: Dispute a lesson completion status
      operationId: disputeCompletion
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: string
          description: Booking ID to dispute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for dispute
              required:
                - reason
      responses:
        '200':
          description: Dispute filed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  dispute_id:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Bookings
  /bookings/:
    get:
      tags:
        - bookings
      summary: List bookings
      description: Get user's bookings with pagination and filtering
      operationId: listBookings
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/BookingStatus'
          description: Filter by booking status
        - in: query
          name: upcoming_only
          schema:
            type: boolean
            default: false
          description: Only show future bookings
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
      responses:
        '200':
          description: Paginated list of bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - bookings
      summary: Create booking
      description: |
        Create a new instant booking.
        Rate limited to 20 per minute per user.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Time slot not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /bookings/upcoming:
    get:
      tags:
        - bookings
      summary: Get upcoming bookings
      description: Get upcoming bookings for dashboard widget
      operationId: getUpcomingBookings
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Number of bookings to return
      responses:
        '200':
          description: List of upcoming bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpcomingBookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /bookings/{booking_id}:
    get:
      tags:
        - bookings
      summary: Get booking details
      description: Get full booking details
      operationId: getBooking
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: integer
          description: Booking ID
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - bookings
      summary: Update booking
      description: Update booking details (instructor only)
      operationId: updateBooking
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: integer
          description: Booking ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /bookings/{booking_id}/complete:
    post:
      tags:
        - bookings
      summary: Complete booking
      description: Mark booking as completed (instructor only)
      operationId: completeBooking
      parameters:
        - in: path
          name: booking_id
          required: true
          schema:
            type: integer
          description: Booking ID
      responses:
        '200':
          description: Booking completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Public Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ServiceBase:
      type: object
      required:
        - skill
        - hourly_rate
      properties:
        skill:
          type: string
          minLength: 1
          maxLength: 100
        hourly_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1000
        description:
          type: string
          maxLength: 500
        duration_override:
          type: integer
          minimum: 30
          maximum: 240
    ServiceResponse:
      allOf:
        - $ref: '#/components/schemas/ServiceBase'
        - type: object
          properties:
            id:
              type: integer
            duration:
              type: integer
              description: Effective duration in minutes
    ServiceAreaNeighborhood:
      type: object
      required:
        - neighborhood_id
        - name
        - borough
      properties:
        neighborhood_id:
          type: string
        ntacode:
          type: string
          nullable: true
        name:
          type: string
        borough:
          type: string
    InstructorProfileBase:
      type: object
      required:
        - bio
        - years_experience
      properties:
        bio:
          type: string
          minLength: 10
          maxLength: 1000
        years_experience:
          type: integer
          minimum: 0
          maximum: 50
        min_advance_booking_hours:
          type: integer
          minimum: 0
          maximum: 168
          default: 2
        buffer_time_minutes:
          type: integer
          minimum: 0
          maximum: 60
          default: 0
...
    InstructorProfileResponse:
      allOf:
        - $ref: '#/components/schemas/InstructorProfileBase'
        - type: object
          properties:
            id:
              type: integer
            user_id:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
              nullable: true
            user:
              type: object
              properties:
                full_name:
                  type: string
                email:
                  type: string
                  format: email
            service_area_neighborhoods:
              type: array
              items:
                $ref: '#/components/schemas/ServiceAreaNeighborhood'
              default: []
            service_area_boroughs:
              type: array
              items:
                type: string
              default: []
            service_area_summary:
              type: string
              nullable: true
            services:
              type: array
              items:
                $ref: '#/components/schemas/ServiceResponse'

    # Availability Schemas
...
    BookingStatus:
      type: string
      enum:
        - PENDING
        - CONFIRMED
        - COMPLETED
        - CANCELLED
        - NO_SHOW
    LocationType:
      type: string
      enum:
        - student_home
        - instructor_location
        - neutral
    BookingCreate:
      type: object
      required:
        - instructor_id
        - service_id
        - booking_date
        - start_time
        - end_time
      properties:
        instructor_id:
          type: integer
        service_id:
          type: integer
        booking_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        student_note:
          type: string
          maxLength: 1000
        meeting_location:
          type: string
        location_type:
          $ref: '#/components/schemas/LocationType'
    BookingUpdate:
      type: object
      properties:
        instructor_note:
          type: string
          maxLength: 1000
        meeting_location:
          type: string
...
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
        student_id:
          type: integer
        instructor_id:
          type: integer
        service_id:
          type: integer
        booking_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        service_name:
          type: string
        hourly_rate:
          type: number
          format: float
        total_price:
          type: number
          format: float
        duration_minutes:
          type: integer
        status:
          $ref: '#/components/schemas/BookingStatus'
        location_type:
          $ref: '#/components/schemas/LocationType'
        meeting_location:
          type: string
          nullable: true
        service_area:
          type: string
          nullable: true
        student_note:
          type: string
          nullable: true
        instructor_note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        student:
          type: object
          properties:
            id:
              type: integer
            full_name:
              type: string
            email:
              type: string
              format: email
        instructor:
          type: object
          properties:
            id:
              type: integer
            full_name:
              type: string
            email:
              type: string
              format: email
        service:
          type: object
          properties:
            id:
              type: integer
            skill:
              type: string
            description:
              type: string
              nullable: true
    BookingListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
    UpcomingBookingResponse:
      type: object
      properties:
        id:
          type: integer
        booking_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        service_name:
          type: string
        student_name:
          type: string
        instructor_name:
          type: string
        meeting_location:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
        field:
          type: string
    RateLimitError:
      type: object
      properties:
        detail:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
              enum:
                - RATE_LIMIT_EXCEEDED
            retry_after:
              type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when the rate limit resets
          schema:
            type: integer